# specs/001-docusaurus-textbook/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Powered Physical AI & Humanoid Robotics Textbook API
  description: |
    Backend API for the interactive textbook platform providing AI-powered features:
    - RAG chatbot with citations
    - Text selection AI explanations
    - Quiz generation and attempt tracking
    - User authentication and progress tracking
    - Content personalization based on user background
    - Urdu translation with RTL support
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: https://api.textbook.example.com/api
    description: Production server
  - url: http://localhost:8000/api
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    ProgressSummary:
      type: object
      properties:
        total_chapters:
          type: integer
        completed_chapters:
          type: integer
        completion_percentage:
          type: number
          format: float
        total_time_minutes:
          type: integer
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/ChapterProgress'
    ChapterProgress:
      type: object
      properties:
        chapter_id:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed]
        quiz_score:
          type: integer
          minimum: 0
          maximum: 100
        time_spent_seconds:
          type: integer
        is_bookmarked:
          type: boolean
        last_accessed:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    UpdateProgressRequest:
      type: object
      properties:
        chapter_id:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed]
        time_spent_seconds:
          type: integer
          minimum: 0

    # Personalization Schemas
...
paths:
...
  /progress:
    get:
      tags:
        - Progress
      summary: Get user's overall progress summary
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Progress summary retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressSummary'
        '401':
          description: Unauthorized
  /progress/{chapter_id}:
    get:
      tags:
        - Progress
      summary: Get progress for a specific chapter
      security:
        - BearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chapter progress retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterProgress'
        '404':
          description: Chapter not found

    put:
      tags:
        - Progress
      summary: Update progress for a chapter
      description: |
        Updates status and time spent. Called when user navigates to/from chapter.
      security:
        - BearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProgressRequest'
      responses:
        '200':
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterProgress'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized

  # Personalization Endpoints
...
security:
  - BearerAuth: []